Deepak -- extern --- fdep.m
Excellent tool for finding program dependencies that i found on matlab central.

http://www.mathworks.com/matlabcentral/fileexchange/17291-fdep-a-pedestrian-function-dependencies-finder

This is apparently 100 times faster that matlab's built-in function depfun which is often times awfully slow even for moderately complex programs.

I have tried it on quite complex programs and it is really fast in comparison to depfun and works like a charm.

I am trying develop a general function that uses this function to create a self-sustained code package for deploying code along with the paper to journals. 


git-svn-id: https://orchestra.med.harvard.edu/svn/lccb/matlab/projects/extern/main@11164 0bcb941b-74d1-0310-bb26-e6636707f451